Logic to display correct text in edit modal

+++ src/components/LoanCard/LoanCard.js
@@ -10,19 +10,35 @@ import EditLoanCard from '../EditLoanCard/EditLoanCard';
 const LoanCard = (props) => {
 
     const deleteBorrower = async (id) => {
-        const response = await fetch(`${API_HOST}/api/v1/borrower/delete/${id}`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        });
-        const data = await response.json();
-        if (!data.errors) {
-            toast.success(data.msg);
-            props.setBorrowers(await getBorrowers());
-        } else {
-            toast.error(data.errors[0].msg);
+        // eslint-disable-next-line
+        const action = confirm('Loan card will be deleted permanently. Are you sure you want to delete?');
+        if (action) {
+            const response = await fetch(`${API_HOST}/api/v1/borrower/delete/${id}`, {
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            });
+            const data = await response.json();
+            if (!data.errors) {
+                toast.success(data.msg);
+                props.setBorrowers(await getBorrowers());
+            } else {
+                toast.error(data.errors[0].msg);
+            }
         }
+    }
+
+    const setEditLoanCardModalTest = () => {
+        const rank = props.borrower.rank;
+        const name = props.borrower.fullName;
+        const serviceNumber = props.borrower.serviceNumber;
+        const department = props.borrower.department;
+
+        document.getElementById(`editRank-${props.borrower._id}`).value = rank;
+        document.getElementById(`editFullName-${props.borrower._id}`).value = name;
+        document.getElementById(`editServiceNumber-${props.borrower._id}`).value = serviceNumber;
+        document.getElementById(`editDepartment-${props.borrower._id}`).value = department;
     }
 
     return (
@@ -35,7 +51,7 @@ const LoanCard = (props) => {
                     <Card.Text>Section: {props.borrower.department}</Card.Text>
                     <div className='controls my-2 d-flex justify-content-end'>
                         <Tooltip title="View"><i className="fas fa-eye mx-1" /></Tooltip>
-                        <Tooltip title="Edit"><i className="fas fa-edit mx-1" data-bs-toggle='modal' data-bs-target={`#edit-borrower-modal-${props.borrower._id}`} /></Tooltip>
+                        <Tooltip title="Edit"><i className="fas fa-edit mx-1" data-bs-toggle='modal' data-bs-target={`#edit-borrower-modal-${props.borrower._id}`} onClick={setEditLoanCardModalTest} /></Tooltip>
                         <Tooltip title="Delete"><i className="fas fa-trash-alt delete-loan-card mx-1" onClick={() => deleteBorrower(props.borrower._id)} /></Tooltip>
                     </div>
                     <Card.Footer><small><b>Created on:</b> {new Date(props.borrower.dataCreationDate).toDateString()}</small></Card.Footer>
